Завершено написание бэкграунда
