+++ Code/Assets/Scripts/IoC/model/ExampleModel.cs
