+++ Code/Assets/Scripts/IoC/model/IExampleModel.cs
