+++ Code/Assets/Scripts/IoC/service/IExampleService.cs
