+++ Code/Assets/Scripts/IoC/view/ClickDetector.cs
