+++ Code/Assets/Scripts/IoC/view/ExampleMediator.cs
