+++ Code/Assets/Scripts/IoC/view/ExampleView.cs
